submissions/python/requirements.txt
pycryptodomex>=3.22.0
ecdsa>=0.18.0